[update move controller under Model namespace
